Pass [CMD] key when [ALT] key is pressed
WhatsApp Web defines its own shortcuts. Some of these shortcuts use
the [ALT] key on Windows and Linux, and the [CMD] key on Mac.
However, since Whatsapp-for-Linux runs on WebKit, WhatsApp Web
listens for the [CMD] key instead of the [ALT] key.

This change causes the [CMD] key to be recieved by the webpage when
the [ALT] key is pressed, so that all standard keyboard shortcuts
provided by WhatsApp Web can be used.

Keyboard shortcuts defined by WhatsApp-for-Linux are not affected.